+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Bug Fixes
 - multiselect - fix bug on chrome tablet where focus was lost from search box
   https://github.com/anvilistas/anvil-extras/pull/627
+- slider - fix padding and scrollbar issues
+  https://github.com/anvilistas/anvil-extras/pull/628
 
 # v3.4.1